# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Benchmarking Service
  version: 0.1.0
paths:
...
  /benchmarks:
    get:
      tags:
      - benchmarks
      summary: List Benchmarks
      operationId: list_benchmarks_benchmarks_get
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response List Benchmarks Benchmarks Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /benchmarks/{name}:
    get:
      tags:
      - benchmarks
      summary: Get Benchmark
      operationId: get_benchmark_benchmarks__name__get
      parameters:
      - name: name
        in: path
        required: true
        schema:
          type: string
          title: Name
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Get Benchmark Benchmarks  Name  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /benchmarks/{name}/deploy:
    post:
      tags:
      - benchmarks
      summary: Deploy Benchmark
      operationId: deploy_benchmark_benchmarks__name__deploy_post
      parameters:
      - name: name
        in: path
        required: true
        schema:
          type: string
          title: Name
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeployBenchmarkRequest'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeployBenchmarkResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - benchmarks
      summary: Teardown Benchmark
      operationId: teardown_benchmark_benchmarks__name__deploy_delete
      parameters:
      - name: name
        in: path
        required: true
        schema:
          type: string
          title: Name
      - name: namespace
        in: query
        required: false
        schema:
          type: string
          default: team1
          title: Namespace
      - name: agent
        in: query
        required: false
        schema:
          type: string
          default: tool_calling
          title: Agent
      - name: experiment
        in: query
        required: false
        schema:
          type: string
          default: default
          title: Experiment
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /benchmarks/{name}/status:
    get:
      tags:
      - benchmarks
      summary: Benchmark Status
      operationId: benchmark_status_benchmarks__name__status_get
      parameters:
      - name: name
        in: path
        required: true
        schema:
          type: string
          title: Name
      - name: namespace
        in: query
        required: false
        schema:
          type: string
          default: team1
          title: Namespace
      - name: agent
        in: query
        required: false
        schema:
          type: string
          default: tool_calling
          title: Agent
      - name: experiment
        in: query
        required: false
        schema:
          type: string
          default: default
          title: Experiment
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BenchmarkStatusResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    BenchmarkStatusResponse:
      properties:
        benchmark:
          type: string
          title: Benchmark
        namespace:
          type: string
          title: Namespace
        tool_name:
          type: string
          title: Tool Name
        agent_name:
          type: string
          title: Agent Name
        tool_ready:
          type: boolean
          title: Tool Ready
        agent_ready:
          type: boolean
          title: Agent Ready
        tool_status:
          anyOf:
          - type: string
          - type: 'null'
          title: Tool Status
        agent_status:
          anyOf:
          - type: string
          - type: 'null'
          title: Agent Status
      type: object
      required:
      - benchmark
      - namespace
      - tool_name
      - agent_name
      - tool_ready
      - agent_ready
      title: BenchmarkStatusResponse
...
    DeployBenchmarkRequest:
      properties:
        agent:
          type: string
          title: Agent
          default: tool_calling
        model:
          anyOf:
          - type: string
          - type: 'null'
          title: Model
        namespace:
          type: string
          title: Namespace
          default: team1
        experiment:
          type: string
          title: Experiment
          default: default
        authbridge_enabled:
          type: boolean
          title: Authbridge Enabled
          default: false
        plugin_preset:
          anyOf:
          - type: string
          - type: 'null'
          title: Plugin Preset
        plugins:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Plugins
        on_error:
          anyOf:
          - type: string
          - type: 'null'
          title: On Error
        plugin_config_file:
          anyOf:
          - type: string
          - type: 'null'
          title: Plugin Config File
      type: object
      title: DeployBenchmarkRequest
    DeployBenchmarkResponse:
      properties:
        benchmark:
          type: string
          title: Benchmark
        namespace:
          type: string
          title: Namespace
        tool_name:
          type: string
          title: Tool Name
        agent_name:
          type: string
          title: Agent Name
        tool:
          additionalProperties: true
          type: object
          title: Tool
        agent:
          additionalProperties: true
          type: object
          title: Agent
      type: object
      required:
      - benchmark
      - namespace
      - tool_name
      - agent_name
      - tool
      - agent
      title: DeployBenchmarkResponse
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Caller JWT. Used only for attribution (preferred_username) and
        issuer-based routing (iss); never forwarded upstream — the Service mints its
        own ROPC token for Rossoctl calls.
security:
- bearerAuth: []